[booster] update tests for booster
