+++ tests/test_checkpoint_io/test_torch_ddp_checkpoint_io.py
@@ -10,7 +10,7 @@ from colossalai.booster import Booster
 from colossalai.booster.plugin import TorchDDPPlugin
 from colossalai.booster.plugin.torch_ddp_plugin import TorchDDPCheckpointIO
 from colossalai.interface import OptimizerWrapper
-from colossalai.testing import clear_cache_before_run, rerun_if_address_is_in_use, spawn
+from colossalai.testing import recursive_check, rerun_if_address_is_in_use, spawn
 
 
 def check_torch_ddp_checkpointIO():
@@ -61,26 +61,3 @@ def run_dist(rank, world_size, port):
 @rerun_if_address_is_in_use()
 def test_torch_ddp_checkpointIO():
     spawn(run_dist, 2)
-
-
-def recursive_check(d1, d2):
-    for k, v in d1.items():
-        if isinstance(v, dict):
-            recursive_check(v, d2[k])
-        elif isinstance(v, list):
-            for i in range(len(v)):
-                if isinstance(v[i], torch.Tensor):
-                    v[i] = v[i].to('cpu')
-                    d2[k][i] = d2[k][i].to('cpu')
-                    if not torch.equal(v[i], d2[k][i]):
-                        raise AssertionError
-                elif v[i] != d2[k][i]:
-                    assert v[i] == d2[k][i]
-
-        elif isinstance(v, torch.Tensor):
-            v = v.to('cpu')
-            d2[k] = d2[k].to('cpu')
-            if not torch.equal(v, d2[k]):
-                raise AssertionError
-        elif not v == d2[k]:
-            raise AssertionError
